backend/requirements.txt
# ============================================================
# Face Recognition Auth System — Python Dependencies
# ============================================================

# --- Web Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# --- Database ---
sqlalchemy==2.0.36
aiosqlite==0.20.0

# --- Authentication ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# --- Encryption ---
cryptography==44.0.0

# --- Face Recognition ---
face-recognition==1.3.0
dlib==19.24.6
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==11.1.0

# --- 2FA ---
pyotp==2.9.0
qrcode[pil]==8.0

# --- Scheduling ---
apscheduler==3.10.4

# --- Environment ---
python-dotenv==1.0.1
pydantic-settings==2.7.1

# --- Testing ---
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# --- Utilities ---
python-dateutil==2.9.0

# --- Voice Biometrics & Deep Learning ---
librosa
soundfile
SpeechRecognition
torch
torchaudio
speechbrain
transformers
scipy
